Ios warning fixes

+++ Source/Framework/MPURLRequestParameter.m
@@ -86,7 +86,7 @@
 + (NSString *)parameterStringForParameters:(NSArray *)inParameters {
 	NSMutableString *queryString = [[NSMutableString alloc] init];
 	int i = 0;
-	int parameterCount = [inParameters count];	
+	unsigned long parameterCount = [inParameters count];
 	MPURLRequestParameter *aParameter = nil;
 	
 	for (; i < parameterCount; i++) {
